Refactor Profile Details into tabs
